Re-organized file structure and completed README
