+++ portfolio_analyzer/future_iteration_framework/controller/analyzer_controller.py
@@ -1,0 +1,2 @@
+# The controller is where the user inputs are captured 
+# The controller delegates data representation to the view and data handling to the model
