+++ portfolio_analyzer/future_iteration_framework/model/analyzer_model.py
@@ -1,0 +1,3 @@
+# backend code for portfolio analyzer
+# business logic of the application
+# access to the database for CRUD operations
