+++ portfolio_analyzer/future_iteration_framework/model/mvc_exceptions.py
@@ -1,0 +1,7 @@
+# mvc_exceptions.py
+class AssetAlreadyStored(Exception):
+    pass
+
+
+class AssetNotStored(Exception):
+    pass
