--- portfolio_analyzer/model/analyzer_model.py
@@ -1,3 +1,0 @@
-# backend code for portfolio analyzer
-# business logic of the application
-# access to the database for CRUD operations
