--- portfolio_analyzer/model/mvc_exceptions.py
@@ -1,7 +1,0 @@
-# mvc_exceptions.py
-class AssetAlreadyStored(Exception):
-    pass
-
-
-class AssetNotStored(Exception):
-    pass
