--- portfolio_analyzer/view/analyzer_view.py
@@ -1,3 +1,0 @@
-# This is the presentation layer.
-# The data is presented to the user
-# The presentation can be in the form of a webpage, a table, chart etc
